pyproject.toml
[tool.poetry]
name = "trueformmanager"
version = "0.1.0"
description = ""
authors = ["sertemo <tejedor.moreno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flet = "^0.19.0"
pymongo = "^4.6.1"
pydantic = "^2.6.1"
pytz = "^2024.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
